Keep runtime-bridge test scaffolding easy to reuse
The runtime bridge tests repeated async collection and model runtime setup inline, so this extracts those helpers into a sibling test-only module without changing assertions or runtime behavior.

Constraint: Refactor must preserve existing runtime-bridge test coverage and behavior
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Move helpers into production runtime code | would expand non-test surface for test-only setup
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep future runtime-bridge test scaffolding in the helper module unless a case is truly one-off
Tested: deno test --no-check --allow-all src/runtime/runtime-bridge.test.ts; deno fmt --check src/runtime/runtime-bridge.test.ts src/runtime/runtime-bridge.test-helpers.ts; deno lint src/runtime/runtime-bridge.test.ts src/runtime/runtime-bridge.test-helpers.ts; git diff --check
Not-tested: Full veryfront-code repo suite
